Added load_levels()
 Moved level1 to file levels.cpp
load_levels returns a list of levels (function pointers). Main then calls each function in a loop.
A new level can be added without changing anyother file.
Each level returns true if the level was passed and false if the player failed the level.

Added << overloads for Rank and Unit